Update package for new classes due to rename
