+++ src/main/java/edu/wpi/first/pathweaver/QuickSpline.java
@@ -1,4 +1,4 @@
-package edu.wpi.first.pathui;
+package edu.wpi.first.pathweaver;
 
 import java.util.Map;
